+++ CHANGELOG.md
@@ -5,15 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+------------------------
+
+## [0.2.0] - 2023-03-06
 ### Added
 - check on 'pip' installation to abort and prevent user when program is not found.
 - command 'doctor' on Linux to detect required tools (only GCC supported for now).
 - command 'config' on Linux to configure custom path instead of PATH environment variable.
 
+### Changed
+- elevated privileges required only when running command 'doctor'.
+
 ### Fixed
 - terminal with a reset of ANSI escape sequences when aborting.
 - check 'python' installation, trying 'python3' as a fallback when 'python' is not found.
-- elevated privileges required only when running command 'doctor'.
 
 ------------------------
 
@@ -68,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 <!-- Table of releases -->
-[Unreleased]: https://github.com/poirierlouis/godot-cli/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/poirierlouis/godot-cli/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/poirierlouis/godot-cli/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/poirierlouis/godot-cli/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/poirierlouis/godot-cli/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/poirierlouis/godot-cli/releases/tag/v0.1.0